Treat 404 errors from Kubernetes API as missing values
Rather than relying on the message returned from the kubernetes api
being exactly "Not Found" we now also treat HTTP code 404 as a
missing configuration value. This is much more robust to changes
as it is unlikely that the HTTP status code spec will change while
kubernetes can decide to change the error message at any point in
time.